Customize cleanup for multi node test
